requirements.txt
backpack-for-pytorch
numpy
scikit_learn
scipy
torch